+++ src/common/extensions.ts
@@ -73,6 +73,15 @@ export interface ToolResult {
 }
 
 /**
+ * Definition of a tool that is part of the tool set. This tool can be executed internally by extension, but it won't be propagated to UI.
+ *
+ * @execute - Optional execute function. If not provided, the tool has other unsupported means of execution.
+ */
+export interface Tool {
+  execute?: (input: Record<string, unknown>) => Promise<unknown>;
+}
+
+/**
  * Definition of a tool that can be registered by an extension
  *
  * @example
@@ -98,7 +107,7 @@ export interface ToolDefinition<TSchema extends z.ZodType = z.ZodType<Record<str
   /** Zod schema for parameter validation */
   inputSchema: TSchema;
   /** Execute function with type-safe args */
-  execute: (input: z.infer<TSchema>, signal: AbortSignal | undefined, context: ExtensionContext) => Promise<unknown>;
+  execute: (input: z.infer<TSchema>, signal: AbortSignal | undefined, context: ExtensionContext, allTools: Record<string, Tool>) => Promise<unknown>;
 }
 
 /** UI element types */